+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.0.0] - 2022-02-05
 ### Added
 - MQTT messages on topic `homeassistant/cover/switchbot-curtain/MAC_ADDRESS/position/set-percent`
   trigger command to set curtain motors' position (payload: decimal integer in range `[0, 100]`)
@@ -117,7 +119,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Subscribe to `homeassistant/switch/switchbot/+/set`.
   Handle `ON` and `OFF` messages.
 
-[Unreleased]: https://github.com/fphammerle/switchbot-mqtt/compare/v2.2.0...HEAD
+[Unreleased]: https://github.com/fphammerle/switchbot-mqtt/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/fphammerle/switchbot-mqtt/compare/v2.2.0...v3.0.0
 [2.2.0]: https://github.com/fphammerle/switchbot-mqtt/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/fphammerle/switchbot-mqtt/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/fphammerle/switchbot-mqtt/compare/v1.1.0...v2.0.0